// RecipeLab/src/enums/enums.tsx
export enum IconName {
  FAVORITEBORDER = 'favorite-border',
  FAVORITE = 'favorite',
  RECIPEICON = 'local-dining',
  ACCOUNTICON = 'account-circle',
  SHOPPINGLISTICON = 'list-alt',
  LOGINICON = 'login',
  SHARE = 'share',
  ADD = 'add',
  REMOVE = 'highlight-off',
  DELETE = 'delete',
  EDIT = 'edit',
  RECIPEADD = 'playlist-add',
}

export enum Colors {
  PRIMARY = '#259909',
  SECONDARY = '#89c87a',
}

export enum Screens {
  SHOPPINGLIST = 'ShoppingList',
  RECIPE = 'Recipe',
  PROFILE = 'Profile',
}